fixing type issues on other systems

+++ common/math.cpp
@@ -67,6 +67,6 @@ uint64_t pareto(uint64_t scale, double shape) {
 
 	std::uniform_real_distribution<> dis(0.0,1.0);
 	double c = pow(dis(gen), 1.0/shape);
-	uint64_t ci = (uint64_t)(((double)(1ul<<32))*c);
-	return min(scale, (scale<<32)/max((uint64_t)1ul,ci) - scale);
+	uint64_t ci = (uint64_t)(((double)((uint64_t)(1)<<32))*c);
+	return min(scale, (scale<<32)/max((uint64_t)(1),ci) - scale);
 }
